fix clang-tidy

+++ tasks/frolova_s_sum_elem_matrix/common/include/common.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <tuple>
 #include <vector>
@@ -10,8 +11,7 @@ namespace frolova_s_sum_elem_matrix {
 
 using Matrix = std::vector<std::vector<int>>;
 using InType = Matrix;
-using OutType = long long;
-// rows, cols, label для функциональных тестов
+using OutType = int64_t;
 using TestType = std::tuple<int, int, std::string>;
 using BaseTask = ppc::task::Task<InType, OutType>;
